HTML, CSS and JavaScript validation form
I added a form to my contact page. I used the edit option in my WordPress website's options and I wrote a custom HTML there. I wrote a validation form function inside my script.js file and tested if it works. I had only one small issue, where I was validating a name and made a tiny mistake inside the nameValidate() function and it was displaying an error message, even if the name was validated correctly. I found the error and removed it. I tested the whole thing again and there were no issues. The form is validating, throwing errors when something is missing or not written properly and clears the form after submission. I styled some elements of my form, added some border-radius to all fields and the button.